+++ CHANGELOG.md
@@ -43,6 +43,7 @@ We use *breaking :warning:* to mark changes that are not backward compatible (re
 
 - [#7334](https://github.com/thanos-io/thanos/pull/7334) Compactor: do not vertically compact downsampled blocks. Such cases are now marked with `no-compact-mark.json`. Fixes panic `panic: unexpected seriesToChunkEncoder lack of iterations`.
 - [#7393](https://github.com/thanos-io/thanos/pull/7393) *: *breaking :warning:* Using native histograms for grpc middleware metrics. Metrics `grpc_client_handling_seconds` and `grpc_server_handling_seconds` will now be native histograms, if you have enabled native histogram scraping you will need to update your PromQL expressions to use the new metric names.
+- [#7472](https://github.com/thanos-io/thanos/pull/7472) Receive: set tenant matched if it matches one of glob patterns.
 
 ### Removed